Linear Regression/requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file handling
numpy>=1.24.0

# Semantic embeddings (local, free)
sentence-transformers>=2.2.0

# Vector search
faiss-cpu>=1.7.4  # Use faiss-gpu if you have GPU

# Optional: For production use
# openai>=1.0.0  # When ready to connect OpenAI API
# python-dotenv>=1.0.0  # For API key management

# Development/testing
pytest>=7.4.0
pytest-cov>=4.1.0
